Whoops forgot to commit in a while hehe
This should contain the flight code for the Due, as well as any other test code that I'd written in support of the mission. The proper Due code should be in the test/due/due.ino file. The pinouts and all my notes are written in the first 45 lines of the file.